[Move] - 캐릭터 정보 파일 폴더 생성 및 이동
