+++ src/app/maple/result/skill/page.tsx
@@ -4,7 +4,7 @@ import React, { Suspense } from "react";
 
 import { Spinner } from "@nextui-org/spinner";
 
-import SkillInfo from "@/components/maple/result/skill/SkillInfo";
+import SkillInfo from "@/components/maple/result/characterInfo/skill/SkillInfo";
 
 const SkillPage = () => {
   return (
