+++ src/components/maple/result/characterInfo/BasicInfo.tsx
