+++ src/components/maple/result/characterInfo/CashItemCard.tsx
