+++ src/components/maple/result/characterInfo/DetailStatInfo.tsx
