+++ src/components/maple/result/characterInfo/InfoNav.tsx
