+++ src/components/maple/result/characterInfo/StatInfo.tsx
