+++ src/components/maple/result/characterInfo/item/ItemCard.tsx
