+++ src/components/maple/result/characterInfo/item/ItemDetail.tsx
