+++ src/components/maple/result/characterInfo/item/ItemList.tsx
