+++ src/components/maple/result/characterInfo/item/ItemPotential.tsx
