+++ src/components/maple/result/characterInfo/item/ItemStat.tsx
