+++ src/components/maple/result/characterInfo/skill/LinkSkillCard.tsx
