+++ src/components/maple/result/characterInfo/skill/LinkSkillDetail.tsx
