+++ src/components/maple/result/characterInfo/skill/LinkSkillList.tsx
