+++ src/components/maple/result/characterInfo/skill/SkillCard.tsx
