+++ src/components/maple/result/characterInfo/skill/SkillDetail.tsx
