+++ src/components/maple/result/characterInfo/skill/SkillInfo.tsx
