+++ src/components/maple/result/characterInfo/symbol/SymbolCard.tsx
