+++ src/components/maple/result/characterInfo/symbol/SymbolDetail.tsx
