+++ src/components/maple/result/characterInfo/symbol/SymbolList.tsx
