+++ src/components/maple/result/characterInfo/union/UnionArtifact.tsx
