+++ src/components/maple/result/characterInfo/union/UnionArtifactCard.tsx
